# 25._OpenAPI/python-OpenAPI/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/api/spacecrafts": {"get": {"tags": ["spacecrafts"], "summary": " ", "operationId": "__api_spacecrafts_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Spacecraft"}, "type": "array", "title": "Response   Api Spacecrafts Get"}}}}}}, "post": {"tags": ["spacecrafts"], "summary": " ", "operationId": "__api_spacecrafts_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Spacecraft"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Spacecraft"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/spacecrafts/{spacecraft_id}": {"get": {"tags": ["spacecrafts"], "summary": " ", "operationId": "__api_spacecrafts__spacecraft_id__get", "parameters": [{"name": "spacecraft_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Spacecraft Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Spacecraft"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Spacecraft": {"properties": {"id": {"type": "integer", "title": "Id"}, "name": {"type": "string", "title": "Name"}}, "type": "object", "required": ["id", "name"], "title": "Spacecraft"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}